Scripts/requirements.txt
pandas
pydriller
gitpython
requests